# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Lotus Driver Tracking System",
    "description": "Backend for driver signup/login, barcode-scan order registration and delivery completion (GPS + timestamp logged per scan), and an SSO-gated admin dashboard for reviewing driver activity.",
    "version": "2.0.0"
  },
  "paths": {
...
    "/api/auth/signup": {
      "post": {
        "summary": "Register a new driver account and issue a login JWT.",
        "operationId": "signup",
        "tags": ["auth"],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/SignupRequest" } } }
        },
        "responses": {
          "201": {
            "description": "Account created.",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/AuthResponse" } } }
          },
          "409": { "description": "An account with this email already exists." },
          "422": { "description": "warehouse_id doesn't exist or is no longer active." }
        }
      }
    },
    "/api/auth/login": {
      "post": {
        "summary": "Log in a driver with email + password and issue a JWT.",
        "operationId": "login",
        "tags": ["auth"],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/LoginRequest" } } }
        },
        "responses": {
          "200": {
            "description": "Login succeeded.",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/AuthResponse" } } }
          },
          "401": { "description": "Invalid email or password." },
          "403": { "description": "This account is disabled." }
        }
      }
    },
    "/api/auth/forgot-password": {
      "post": {
        "summary": "Request a password-reset link by email; always returns a generic response so the endpoint can't be used to enumerate accounts.",
        "operationId": "forgotPassword",
        "tags": ["auth"],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ForgotPasswordRequest" } } }
        },
        "responses": {
          "200": {
            "description": "Always returned, whether or not the email has an account.",
            "content": {
              "application/json": {
                "schema": { "type": "object", "properties": { "message": { "type": "string" } } }
              }
            }
          }
        }
      }
    },
    "/api/auth/reset-password": {
      "post": {
        "summary": "Complete a password reset using the token from the emailed link, and log the driver in.",
        "operationId": "resetPassword",
        "tags": ["auth"],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ResetPasswordRequest" } } }
        },
        "responses": {
          "200": {
            "description": "Password updated; a fresh login JWT is issued.",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/AuthResponse" } } }
          },
          "400": { "description": "The reset link is invalid, already used, or expired." }
        }
      }
    },
...
    "/api/statuses": {
      "get": {
        "summary": "List the active job status codes (registered, delivered, cancelled).",
        "operationId": "listStatuses",
        "tags": ["common"],
        "security": [{ "driverBearer": [] }, { "adminSSO": [] }],
        "responses": {
          "200": {
            "description": "Status list, in display order.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": { "statuses": { "type": "array", "items": { "$ref": "#/components/schemas/Status" } } }
                }
              }
            }
          }
        }
      }
    },
    "/api/photos/{photo_id}": {
      "get": {
        "summary": "Fetch a stored proof photo's raw bytes; drivers may only fetch their own uploads, admins may fetch any.",
        "operationId": "getPhoto",
        "tags": ["common"],
        "security": [{ "driverBearer": [] }, { "adminSSO": [] }],
        "parameters": [{ "name": "photo_id", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "responses": {
          "200": { "description": "Raw image bytes.", "content": { "image/jpeg": { "schema": { "type": "string", "format": "binary" } } } },
          "403": { "description": "Not your photo." },
          "404": { "description": "Photo not found." }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "driverBearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Driver JWT from /api/auth/signup or /api/auth/login, signed with the platform's JWT_SECRET."
      },
      "adminSSO": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Forwarded-Email",
        "description": "Injected by the Substrait platform's Google SSO proxy when SSO is enabled for this app in the portal's Access tab; not client-supplied. The caller must also be on the admin allowlist in the users table."
      }
    },
    "schemas": {
      "SignupRequest": {
        "type": "object",
        "required": ["email", "password", "name"],
        "properties": {
          "email": { "type": "string" },
          "password": { "type": "string" },
          "name": { "type": "string" },
          "phone": { "type": "string", "nullable": true },
          "warehouse_id": { "type": "integer", "nullable": true, "description": "The driver's assigned outlet, from GET /api/warehouses." }
        }
      },
      "LoginRequest": {
        "type": "object",
        "required": ["email", "password"],
        "properties": { "email": { "type": "string" }, "password": { "type": "string" } }
      },
      "ForgotPasswordRequest": {
        "type": "object",
        "required": ["email"],
        "properties": { "email": { "type": "string" } }
      },
      "ResetPasswordRequest": {
        "type": "object",
        "required": ["token", "new_password"],
        "properties": { "token": { "type": "string" }, "new_password": { "type": "string" } }
      },
...
      "AuthResponse": {
        "type": "object",
        "properties": {
          "token": { "type": "string" },
          "user": {
            "type": "object",
            "properties": {
              "id": { "type": "integer" },
              "email": { "type": "string" },
              "name": { "type": "string" },
              "warehouse_id": { "type": "integer", "nullable": true },
              "warehouse_name": { "type": "string", "nullable": true }
            }
          }
        }
      },
...
      "Status": {
        "type": "object",
        "properties": {
          "code": { "type": "string" },
          "label": { "type": "string" },
          "is_terminal_success": { "type": "boolean" }
        }
      },
...
    }
  }
}